fix build

+++ cumulus/client/consensus/aura/src/collators/slot_based/mod.rs
@@ -256,7 +256,7 @@ struct CollatorMessage<Block: BlockT> {
 	/// Core index that this block should be submitted on
 	pub core_index: CoreIndex,
 	/// Maximum pov size. Currently needed only for exporting PoV.
-	pub max_pov_size: usize,
+	pub max_pov_size: u32,
 }
 
 /// Fetch the `CoreSelector` and `ClaimQueueOffset` for `parent_hash`.
